fix: 지원자의 질문별 점수 계산 방식 변경
- 단순 점수 총점이 아니라 면접관 수(평가 수)로 나눈 값을 반환해야 한다. 또한 중요도(가중치)의 범위는 3~5이다.

+++ src/main/java/com/gotcha/server/evaluation/dto/response/QuestionRankResponse.java
@@ -1,5 +1,5 @@
 package com.gotcha.server.evaluation.dto.response;
 
-public record QuestionRankResponse(Long id, Integer totalScore) {
+public record QuestionRankResponse(Long id, Double totalScore) {
 
 }
